Modified visualize.py
